+++ database/migrations/301_create_workorder_attachments.sql
@@ -1,14 +1,14 @@
 -- Migration 301 — Work Order Attachments (§12)
 CREATE TABLE IF NOT EXISTS work_order_attachments (
-  id               INT UNSIGNED NOT NULL AUTO_INCREMENT,
-  work_order_id    INT UNSIGNED NOT NULL,
+  id               BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
+  work_order_id    BIGINT UNSIGNED NOT NULL,
   filename         VARCHAR(255) NOT NULL,
   original_filename VARCHAR(255) NOT NULL,
   mime_type        VARCHAR(100) NOT NULL,
   file_size        INT UNSIGNED NOT NULL,
   storage_path     VARCHAR(500) NOT NULL,
-  uploaded_by      INT UNSIGNED NOT NULL,
-  organization_id  INT UNSIGNED NOT NULL,
+  uploaded_by      BIGINT UNSIGNED NOT NULL,
+  organization_id  BIGINT UNSIGNED NOT NULL,
   created_at       TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (id),
   KEY idx_wo_attachments_work_order (work_order_id),
